// src/app/interfaces/settings.ts
export class Translations{
    headings: {
      inventory: String;
      conf_notif: String;
      printing_option: String;
      multi_box_option: String;
      per_box_packing_slip: String;
      miscellaneous_settings: String;
      flash_setting: String;
      time_zone: String;
      adjust_for_daylight: String;
      order_item_export: String;
      order_tracking_no_order: String;
      order_tracking_no_tracking: String;
      product_weight: String;
      simple_product: String;
      pre_day_order: String;
      pre_day_import: String;
      import_orders: String;
      recurring_ftp: String
    };
    labels: {
      inventory_tracking: String;
      low_inventory_alert_email: String;
      time_to_send_email: String;
      send_email_on: String;
      mon: String;
      tue: String;
      wed: String;
      thu: String;
      fri: String;
      sat: String;
      sun: String;
      strict_cc: String;
      conf_code_product_instruction: String;
      default_low_inventory_alert_limit: String;
      conf_req_on_notes_to_packer: String;
      send_email_for_packer_notes: String;
      email_address: String;
      system_notifications: String;
      billing_notifications: String;
      packing_slip_size: String;
      packing_slip_orientation: String;
      portrait: String;
      landscape: String;
      packing_slip_message_to_customer: String;
      show_primary_bin_loc_in_barcodeslip: String;
      inventory_auto_allocation: String;
      custom_field_one: String;
      custom_field_two: String;
      html_print: String;
      inactive_logout_time: String;
      hex_barcode: String;
      multi_box_shipment: String;
      custom_user_field_one: String;
      custom_user_field_two: String;
      display_kit_parts: String;
      remove_order_items: String;
      print_when_order_complete: String;
      print_when_new_box: String;
      manually_print: String;
      display_sku_total: String;
      display_one_total: String
    };
    tooltips: {
      inventory_tracking: String;
      low_inventory_alert_email: String;
      default_low_inventory_alert_limit: String;
      conf_req_on_notes_to_packer: String;
      strict_cc: String;
      conf_code_product_instruction: String;
      send_email_for_packer_notes: String;
      packing_slip_size: String;
      packing_slip_message_to_customer: String;
      show_primary_bin_loc_in_barcodeslip: String;
      order_item_export_message: String;
      simple_product_scanning: String;
      html_print: String;
      inactive_logout_time: String;
      hex_barcode: String;
      multi_box_shipment: String;
      custom_fields: String;
      pre_day_order_import_schedule: String;
      recurring_ftp_order_import_range: String;
      display_kit_parts: String;
      remove_order_items: String
    }
  };